+++ TIUIElements/Sources/Views/StatefulButton/RoundedStatefulButton.swift
@@ -20,6 +20,7 @@
 //  THE SOFTWARE.
 //
 
+import TIUIKitCore
 import UIKit
 
 open class RoundedStatefulButton: StatefulButton {